refactor: remove unused Rest class
