+++ packages/client/src/StorageNodeRegistry.ts
@@ -10,7 +10,7 @@ import { BrubeckContainer } from './Container'
 import { ConfigInjectionToken, StrictStreamrClientConfig } from './Config'
 import { Stream, StreamProperties } from './Stream'
 import { Ethereum } from './Ethereum'
-import { NotFoundError } from './authFetch'
+import { NotFoundError } from './fetchStream'
 import { EthereumAddress, StreamID, toStreamID } from 'streamr-client-protocol'
 import { StreamIDBuilder } from './StreamIDBuilder'
 import { waitForTx, withErrorHandlingAndLogging } from './utils/contract'
