+++ packages/client/src/index-exports.ts
@@ -54,7 +54,7 @@ export { UpdateEncryptionKeyOptions } from './encryption/GroupKeyStoreFactory'
 
 export { ConfigTest } from './ConfigTest'
 export { NetworkNodeStub } from './BrubeckNode'
-export { NotFoundError, ErrorCode } from './authFetch'
+export { NotFoundError, ErrorCode } from './fetchStream'
 export { SignalListener } from './utils/Signal'
 export * from './types'
 export { formStorageNodeAssignmentStreamId } from './utils/index'
